feat: Create SearchCard Component
