+++ src/pages/Search/index.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from "react";
 import { instance } from "../../api/api";
 import { useLocation } from "react-router-dom";
-import styled from "styled-components";
-import SearchDetail from "../../components/SearchDetail";
+import SearchCard from "../../components/SearchCard";
 
 const Search = () => {
   const [result, setResult] = useState<any>([]);
@@ -33,95 +32,11 @@ const Search = () => {
 
   return (
     <main>
-      {result.map((data: any, i: number) => {
-        return (
-          <Section key={i}>
-            <Card>
-              <Video>
-                <img src={data.snippet.thumbnails.medium.url} alt="video" />
-                <SearchDetail
-                  videoId={data.id.videoId}
-                  duration="duration"
-                ></SearchDetail>
-              </Video>
-            </Card>
-            <Info>
-              <Title>{data.snippet.title}</Title>
-              <Views>
-                <SearchDetail
-                  videoId={data.id.videoId}
-                  views="views"
-                ></SearchDetail>{" "}
-                <SearchDetail
-                  videoId={data.id.videoId}
-                  time="time"
-                ></SearchDetail>
-              </Views>
-              <Name>
-                <button></button>
-                <span>{data.snippet.channelTitle}</span>
-              </Name>
-              <Descript>{data.snippet.description}</Descript>
-            </Info>
-          </Section>
-        );
+      {result.map((data: {}, i: number) => {
+        return <SearchCard data={data} key={i}></SearchCard>;
       })}
     </main>
   );
 };
-
-const Section = styled.section`
-  padding: 1.5rem 2rem;
-  display: flex;
-`;
-
-const Card = styled.div`
-  height: 200px;
-  color: #fff;
-  font-size: 15px;
-  letter-spacing: 0.2px;
-`;
-
-const Video = styled.div`
-  max-width: 360px;
-  height: 200px;
-  position: relative;
-  margin-right: 20px;
-  img {
-    border-radius: 10px;
-    height: 100%;
-  }
-`;
-
-const Info = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const Title = styled.h3`
-  margin-bottom: 0.5rem;
-  line-height: 150%;
-  font-size: 18px;
-`;
-
-const Views = styled.div`
-  font-size: 12px;
-  color: rgb(96, 96, 96);
-`;
-
-const Name = styled(Views)`
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  padding: 12px 0;
-  button {
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    border: none;
-  }
-`;
-
-const Descript = styled(Views)``;
 
 export default Search;
